+++ seqkit/cmd/helper.go
@@ -265,16 +265,54 @@ func sortRecordChunkMapID(chunks map[uint64]fastx.RecordChunk) sortutil.Uint64Sl
 }
 
 func filepathTrimExtension(file string) (string, string) {
-	gz := strings.HasSuffix(file, ".gz") || strings.HasSuffix(file, ".GZ")
-	if gz {
+	var _gz, _xz, _zst bool
+	_gz = strings.HasSuffix(file, ".gz") || strings.HasSuffix(file, ".GZ")
+	if _gz {
 		file = file[0 : len(file)-3]
+	} else {
+		_xz = strings.HasSuffix(file, ".xz") || strings.HasSuffix(file, ".XZ")
+		if _xz {
+			file = file[0 : len(file)-3]
+		} else {
+			_zst = strings.HasSuffix(file, ".zst") || strings.HasSuffix(file, ".ZST")
+			if _zst {
+				file = file[0 : len(file)-4]
+			}
+		}
 	}
+
 	extension := filepath.Ext(file)
 	name := file[0 : len(file)-len(extension)]
-	if gz {
+	if _gz {
 		extension += ".gz"
+	} else if _xz {
+		extension += ".xz"
+	} else if _zst {
+		extension += ".zst"
 	}
 	return name, extension
+}
+
+func filepathTrimExtension2(file string, suffixes []string) (string, string, string) {
+	if suffixes == nil {
+		suffixes = []string{".gz", ".xz", ".zst"}
+	}
+
+	var e, e1, e2 string
+	f := strings.ToLower(file)
+	for _, s := range suffixes {
+		e = strings.ToLower(s)
+		if strings.HasSuffix(f, e) {
+			e2 = e
+			file = file[0 : len(file)-len(e)]
+			break
+		}
+	}
+
+	e1 = filepath.Ext(file)
+	name := file[0 : len(file)-len(e1)]
+
+	return name, e1, e2
 }
 
 func isPlainFile(file string) bool {